Added new buttons for History and graphs activities, new activity history(in progress).
I added two buttons to change to new activities. New activity history has been added. Currently is able to show in ListView content of readingBuffer (about all of the reading that he gathered on the mainScreen). Also, when switching the activity to history, BT is stopped (gotta fix that).I also realized that milis is not being saved in buffer (gotta take a look at that). And possibly change three ListViews for one edited.